+++ src/views/seq/product.vue
@@ -6,7 +6,7 @@
       <div class="table-search">
         <a-form layout="inline" :form="form" @submit.prevent="handleSearch">
           <a-row :gutter="24">
-            <a-col :xxl="4" :xl="6" :md="8">
+            <a-col :md="6" :xl="4">
               <a-form-item label="状态">
                 <a-select v-decorator="['status', {initialValue: 1}]">
                   <a-select-option value="">全部</a-select-option>
@@ -14,12 +14,12 @@
                 </a-select>
               </a-form-item>
             </a-col>
-            <a-col :xxl="4" :xl="6" :md="8">
+            <a-col :md="6" :xl="4">
               <a-form-item label="SAP产品编号">
                 <a-input-search v-decorator="['productCode']" @search="searchProduct"/>
               </a-form-item>
             </a-col>
-            <a-col :xxl="4" :xl="6" :md="8">
+            <a-col :md="6" :xl="4">
               <a-form-item label="样品类型">
                 <a-select v-decorator="['sampleTypeId', {initialValue: ''}]">
                   <a-select-option value="">全部</a-select-option>
@@ -27,7 +27,7 @@
                 </a-select>
               </a-form-item>
             </a-col>
-            <a-col :xxl="4" :xl="6" :md="8">
+            <a-col :md="6" :xl="4">
               <a-form-item label="测序类型">
                 <a-select v-decorator="['seqTypeId', {initialValue: ''}]">
                   <a-select-option value="">全部</a-select-option>
@@ -50,14 +50,15 @@
       <vxe-grid
         highlight-hover-row
         auto-resize
-        :ref="productTable.ref"
+        height="600"
+        ref="productTable"
         :loading="productTable.loading"
         :columns="productTable.columns"
         :pager-config="productTable.pagerConfig"
         :data.sync="productTable.tableData"
         :edit-rules="productTable.editRules"
         :edit-config="{key: 'id', trigger: 'manual', mode: 'row', showIcon: false, autoClear: false}"
-        @page-change="pagerChange">
+        @page-change="({pageSize, currentPage}) => this.$utils.tablePageChange({pageSize, currentPage, table: productTable, callback: handleSearch})">
       </vxe-grid>
     </div>
   </div>
@@ -72,10 +73,7 @@ export default {
       queryParam: {},
       productTable: {
         id: 0,
-        ref: 'productTable',
         xTable: null,
-        editIndex: -1,
-        editData: null,
         loading: false,
         tableData: [],
         columns: [],
@@ -148,14 +146,14 @@ export default {
 
               if (!isEdit && row.status === 1) {
                 actions = [
-                  <a onClick={() => this.handleCancel(options)}>删除</a>,
-                  <a onClick={() => this.handleUpdate(options)}>修改</a>
+                  <a onClick={ () => this.handleCancel(options) }>删除</a>,
+                  <a onClick={ () => xTable.setActiveRow(row) }>修改</a>
                 ];
               }
               if (isEdit) {
                 actions = [
-                  <a onClick={() => this.handleSave(options) }>保存</a>,
-                  <a onClick={() => this.handleQuitEdit(options) }>退出</a>
+                  <a onClick={ () => this.handleSave(options) }>保存</a>,
+                  <a onClick={ () => this.$utils.tableQuitEdit(options) }>退出</a>
                 ];
               }
 
@@ -175,15 +173,18 @@ export default {
 
       this[tableName].columns = columns;
       this[tableName].editRules = {
-        name: [
-          { required: true, message: '名称不能为空' }
+        productName: [
+          { required: true, message: 'SAP产品不能为空' }
         ],
-        seriesId: [
-          { required: true, message: '系列不能为空' }
+        sampleTypeId: [
+          { required: true, message: '样品类型不能为空' }
+        ],
+        seqTypeId: [
+          { required: true, message: '测序类型不能为空' }
         ]
       };
 
-      this[tableName].xTable = this.$refs[this[tableName].ref].$refs.xTable;
+      this[tableName].xTable = this.$refs[tableName].$refs.xTable;
     },
     // 查询
     handleSearch () {
@@ -199,8 +200,6 @@ export default {
         this[tableName].pagerConfig.total = data.total;
         this[tableName].pagerConfig.currentPage = params.page;
         this[tableName].pagerConfig.pageSize = params.rows;
-
-        this[tableName].editIndex = -1;
       }).finally(() => {
         this[tableName].loading = false;
       });
@@ -208,22 +207,18 @@ export default {
     // 新增一可编辑行
     handleAddRow () {
       const tableName = 'productTable';
-      if (this[tableName].editIndex !== -1) return this.$message.warning('请保存或退出正在编辑的行');
-
       const table = this[tableName].xTable;
+
+      const active = table.getActiveRow();
+      if (active && active.row) return this.$message.warning('请保存或退出正在编辑的行');
+
       const newData = {
         id: --this[tableName].id
       };
 
-      this[tableName].tableData = [newData, ...this[tableName].tableData];
-      table.setActiveRow(newData);
-      this[tableName].editIndex = 0;
-    },
-    // 修改
-    handleUpdate ({ row, rowIndex, tableName, xTable }) {
-      xTable.setActiveRow(row);
-      this[tableName].editIndex = rowIndex;
-      this[tableName].editData = JSON.parse(JSON.stringify(row));
+      table.insert(newData).then(({ row }) => {
+        table.setActiveRow(row);
+      });
     },
     // 删除
     handleCancel ({ row }) {
@@ -232,33 +227,15 @@ export default {
       });
     },
     // 保存
-    handleSave ({ row }) {
-      if (row.status) {
-        // 修改(没有修改功能)
-      } else {
+    handleSave ({ row, xTable }) {
+      xTable.validate(row).then(() => {
         // 新增
         this.$api.sampletype.addSeqProduct(row).then(() => {
           this.handleSearch();
         });
-      }
-    },
-    // 退出编辑
-    handleQuitEdit ({ row, rowIndex, tableName, xTable }) {
-      xTable.clearActived().then(() => {
-        this[tableName].editIndex = -1;
-        if (!row.status) {
-          this[tableName].tableData.splice(rowIndex, 1);
-        } else {
-          this.$set(this[tableName].tableData, rowIndex, this[tableName].editData);
-          this[tableName].editData = null;
-        }
+      }).catch(err => {
+        console.log(err);
       });
-    },
-    // 分页改变时
-    pagerChange ({ pageSize, currentPage }) {
-      const tableName = 'productTable';
-      this[tableName].pagerConfig = Object.assign(this[tableName].pagerConfig, { pageSize, currentPage });
-      this.handleSearch();
     },
     // 选择样品类型
     sampleTypeChange (arr, row, value) {