+++ src/views/seq/series.vue
@@ -5,17 +5,17 @@
     <div class="table-search">
       <a-form layout="inline" :form="form" @submit.prevent="handleSearch">
         <a-row :gutter="24">
-          <a-col :xxl="4" :xl="6" :md="8">
+          <a-col :md="6" :xl="4">
             <a-form-item label="编号">
               <a-input v-decorator="['code']"/>
             </a-form-item>
           </a-col>
-          <a-col :xxl="4" :xl="6" :md="8">
+          <a-col :md="6" :xl="4">
             <a-form-item label="名称">
               <a-input v-decorator="['name']"/>
             </a-form-item>
           </a-col>
-          <a-col :xxl="4" :xl="6" :md="8">
+          <a-col :md="6" :xl="4">
             <a-form-item label="状态">
               <a-select v-decorator="['status', {initialValue: 1}]">
                 <a-select-option value="">全部</a-select-option>
@@ -43,15 +43,15 @@
           highlight-hover-row
           auto-resize
           height="570"
-          :ref="seriesTable.ref"
+          ref="seriesTable"
           :loading="seriesTable.loading"
           :columns="seriesTable.columns"
           :pager-config="seriesTable.pagerConfig"
           :data.sync="seriesTable.tableData"
           :edit-rules="seriesTable.editRules"
           :edit-config="{key: 'id', trigger: 'manual', mode: 'row', showIcon: false, autoClear: false}"
-          @cell-click="(options) => handleCellClick(options)"
-          @page-change="pagerChange">
+          @cell-click="({row}) => handleSearchPrimer(row.id)"
+          @page-change="({pageSize, currentPage}) => this.$utils.tablePageChange({pageSize, currentPage, table: seriesTable, callback: handleSearch})">
         </vxe-grid>
       </a-layout-content>
 
@@ -67,7 +67,7 @@
           highlight-hover-row
           auto-resize
           height="570"
-          :ref="seriesPrimersTable.ref"
+          ref="seriesPrimersTable"
           :loading="seriesPrimersTable.loading"
           :columns="seriesPrimersTable.columns"
           :data.sync="seriesPrimersTable.tableData"
@@ -83,7 +83,7 @@
       :visible="choosePrimer.visible"
       :footer="null"
       @cancel="choosePrimer.visible = false">
-      <choose-primer @callback="setPrimer"></choose-primer>
+      <choose-primer @callback="setPrimerBySelect"></choose-primer>
     </a-modal>
   </div>
 </template>
@@ -92,7 +92,7 @@
 import ChoosePrimer from '@/components/seq/choosePrimer';
 
 export default {
-  name: 'Series',
+  name: 'SeqSeries',
   components: {
     ChoosePrimer
   },
@@ -103,10 +103,7 @@ export default {
       seriesTable: {
         child: 'seriesPrimersTable',
         id: 0,
-        ref: 'seriesTable',
         xTable: null,
-        editIndex: -1,
-        editData: null,
         loading: false,
         tableData: [],
         columns: [],
@@ -115,19 +112,12 @@ export default {
           pageSize: 10,
           total: 0
         },
-        editRules: {
-          name: [
-            { required: true, message: '名称必填' }
-          ]
-        }
+        editRules: {}
       },
       seriesPrimersTable: {
         parent: 'seriesTable',
         id: 0,
-        ref: 'seriesPrimersTable',
         xTable: null,
-        editIndex: -1,
-        editData: null,
         loading: false,
         tableData: [],
         columns: [],
@@ -152,11 +142,10 @@ export default {
       const { basic } = this.$store.state;
 
       const columns = [
-        // { type: 'radio', width: 40 },
-        { type: 'selection', width: 40 },
+        { type: 'radio', width: 40 },
         { type: 'index', width: 40 },
         { title: '编号', field: 'code' },
-        { title: '名称', field: 'name', editRender: { name: 'AInput' } },
+        { title: '名称', field: 'name', editRender: { name: 'input' } },
         { title: '状态', field: 'status', formatter: function ({ cellValue }) { return formatter(basic.status, cellValue); } },
         { title: '创建人', field: 'creatorName' },
         { title: '创建时间', field: 'createDate' },
@@ -177,14 +166,14 @@ export default {
 
               if (!isEdit && row.status === 1) {
                 actions = [
-                  <a onClick={() => this.handleCancel(options)}>删除</a>,
-                  <a onClick={() => this.handleUpdate(options)}>修改</a>
+                  <a onClick={ () => this.handleCancel(options) }>删除</a>,
+                  <a onClick={ () => xTable.setActiveRow(row) }>修改</a>
                 ];
               }
               if (isEdit) {
                 actions = [
-                  <a onClick={() => this.handleSave(options) }>保存</a>,
-                  <a onClick={() => this.handleQuitEdit(options) }>退出</a>
+                  <a onClick={ () => this.handleSave(options) }>保存</a>,
+                  <a onClick={ () => this.$utils.tableQuitEdit(options) }>退出</a>
                 ];
               }
 
@@ -203,8 +192,13 @@ export default {
       });
 
       this[tableName].columns = columns;
+      this[tableName].editRules = {
+        name: [
+          { required: true, message: '名称必填' }
+        ]
+      };
 
-      this[tableName].xTable = this.$refs[this[tableName].ref].$refs.xTable;
+      this[tableName].xTable = this.$refs[tableName].$refs.xTable;
     },
     // 查询
     handleSearch () {
@@ -221,7 +215,8 @@ export default {
         this[tableName].pagerConfig.currentPage = params.page;
         this[tableName].pagerConfig.pageSize = params.rows;
 
-        // this[tableName].editIndex = -1;
+        // 重置子列表
+        this.handleSearchPrimer();
       }).finally(() => {
         this[tableName].loading = false;
       });
@@ -229,25 +224,18 @@ export default {
     // 新增一可编辑行
     handleAddRow () {
       const tableName = 'seriesTable';
-      if (this[tableName].editIndex !== -1) return this.$message.warning('请保存或退出正在编辑的行');
-
       const table = this[tableName].xTable;
+
+      const active = table.getActiveRow();
+      if (active && active.row) return this.$message.warning('请保存或退出正在编辑的行');
+
       const newData = {
         id: --this[tableName].id
       };
 
-      // this[tableName].tableData = [newData, ...this[tableName].tableData];
-      // table.setActiveRow(newData);
       table.insert(newData).then(({ row }) => {
         table.setActiveRow(row);
       });
-      this[tableName].editIndex = 0;
-    },
-    // 修改
-    handleUpdate ({ row, rowIndex, tableName, xTable }) {
-      xTable.setActiveRow(row);
-      this[tableName].editIndex = rowIndex;
-      this[tableName].editData = JSON.parse(JSON.stringify(row));
     },
     // 删除
     handleCancel ({ row }) {
@@ -256,53 +244,22 @@ export default {
       });
     },
     // 保存
-    handleSave ({ row }) {
-      if (row.status) {
-        // 修改
-        this.$api.series.updateSeries(row).then(() => {
-          this.handleSearch();
-        });
-      } else {
-        // 新增
-        this.$api.series.addSeries(row).then(() => {
-          this.handleSearch();
-        });
-      }
-    },
-    /**
-     * 退出编辑
-     * status 有值，退出修改操作，还原数据
-     *        无值，退出新增操作，删除新增的行
-     */
-    handleQuitEdit ({ row, xTable }) {
-      xTable.clearActived().then(() => {
-        if (typeof row.status === 'number') {
-          xTable.revert(row);
+    handleSave ({ row, xTable }) {
+      xTable.validate(row).then(() => {
+        if (row.status) {
+          // 修改
+          this.$api.series.updateSeries(row).then(() => {
+            this.handleSearch();
+          });
         } else {
-          xTable.remove(row);
+          // 新增
+          this.$api.series.addSeries(row).then(() => {
+            this.handleSearch();
+          });
         }
+      }).catch(err => {
+        console.log(err);
       });
-    },
-    // 点击表格行时
-    handleCellClick ({ row }) {
-      const tableName = 'seriesPrimersTable';
-      if (!row.id || row.id < 0) {
-        this[tableName].tableData = [];
-        return;
-      }
-
-      this[tableName].loading = true;
-      this.$api.series.getPrimersBySeries(row.id).then(res => {
-        this[tableName].tableData = res;
-      }).finally(() => {
-        this[tableName].loading = false;
-      });
-    },
-    // 分页改变时
-    pagerChange ({ pageSize, currentPage }) {
-      const tableName = 'seriesTable';
-      this[tableName].pagerConfig = Object.assign(this[tableName].pagerConfig, { pageSize, currentPage });
-      this.handleSearch();
     },
 
     /**
@@ -328,12 +285,12 @@ export default {
               const options = { row, rowIndex, tableName, xTable };
               if (!isEdit) {
                 actions = [
-                  <a onClick={() => this.handleCancel(options)}>删除</a>
+                  <a onClick={ () => this.handleCancelPrimer(options) }>删除</a>
                 ];
               } else {
                 actions = [
-                  <a onClick={() => this.handleSave(options) }>保存</a>,
-                  <a onClick={() => this.handleQuitEdit(options) }>退出</a>
+                  <a onClick={ () => this.handleSavePrimer(options) }>保存</a>,
+                  <a onClick={ () => this.$utils.tableQuitEdit(options) }>退出</a>
                 ];
               }
               return [
@@ -347,29 +304,55 @@ export default {
       ];
 
       this[tableName].columns = columns;
+      this[tableName].editRules = {
+        name: [
+          { required: true, message: '名称必填' }
+        ]
+      };
 
-      this[tableName].xTable = this.$refs[this[tableName].ref].$refs.xTable;
+      this[tableName].xTable = this.$refs[tableName].$refs.xTable;
+    },
+    // 查询
+    handleSearchPrimer (seriesId) {
+      const tableName = 'seriesPrimersTable';
+
+      if (!seriesId || seriesId < 0) {
+        this[tableName].tableData = [];
+        return;
+      }
+
+      this[tableName].loading = true;
+      this.$api.series.getPrimersBySeries(seriesId).then(res => {
+        this[tableName].tableData = res;
+      }).finally(() => {
+        this[tableName].loading = false;
+      });
     },
     // 新增一可编辑行
     handleAddRowToPrimers () {
       const tableName = 'seriesPrimersTable';
       const table = this[tableName].xTable;
       const parentTable = this[this[tableName].parent].xTable;
-      console.log(parentTable.getSelectRecords());
+      const selectRow = parentTable.getCurrentRow();
+      if (!selectRow) return this.$message.warning('请先选择左侧列表的一行');
+      if (selectRow.status !== 1) return this.$message.warning('只能为状态正常的系列添加引物');
+      const active = table.getActiveRow();
+      if (active && active.row) return this.$message.warning('请保存或退出正在编辑的行');
+
       const newData = {
+        seriesId: selectRow.id,
         id: --this[tableName].id
       };
       table.insert(newData).then(({ row }) => {
         table.setActiveRow(row);
       });
-      // this[tableName].editIndex = 0;
     },
     // 显示引物选择框
     selectPrimer () {
       this.choosePrimer.visible = true;
     },
     // 设置引物
-    setPrimer (data) {
+    setPrimerBySelect (data) {
       const tableName = 'seriesPrimersTable';
       const table = this[tableName].xTable;
       const primer = {
@@ -381,6 +364,22 @@ export default {
       Object.assign(table.getInsertRecords()[0], primer);
 
       this.choosePrimer.visible = false;
+    },
+    // 保存引物
+    handleSavePrimer ({ row, xTable }) {
+      xTable.validate(row).then(() => {
+        this.$api.series.addPrimersBySeries(row).then(() => {
+          this.handleSearchPrimer(row.seriesId);
+        });
+      }).catch(err => {
+        console.log(err);
+      });
+    },
+    // 删除引物
+    handleCancelPrimer ({ row }) {
+      this.$api.series.cancelPrimersBySeries(row).then(() => {
+        this.handleSearchPrimer(row.seriesId);
+      });
     }
   }
 };